+++ src/app/loading.tsx
@@ -1,0 +1,3 @@
+export default function Loading() {
+  return "Loading..." // TODO: add something cool
+}
