+++ src/components/TodoItem.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 import { FC } from "react"
 
 const TodoItem: FC = () => {
