+++ src/utils/prisma.ts
@@ -1,3 +1,7 @@
 import { PrismaClient } from "@prisma/client"
 
 export const prisma = new PrismaClient()
+
+process.on("beforeExit", () => {
+  prisma.$disconnect()
+})
